# Conflicts:
#	SimulationConstructionSetTools/src/us/ihmc/simulationConstructionSetTools/util/environments/HatchEnvironment.java

+++ IHMCAvatarInterfaces/test/us/ihmc/avatar/testTools/DRCSimulationTestHelper.java
@@ -29,6 +29,7 @@ import us.ihmc.communication.net.PacketConsumer;
 import us.ihmc.communication.packetCommunicator.PacketCommunicator;
 import us.ihmc.communication.packets.Packet;
 import us.ihmc.communication.util.NetworkPorts;
+import us.ihmc.euclid.geometry.BoundingBox3D;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.Vector3D;
 import us.ihmc.humanoidBehaviors.behaviors.scripts.engine.ScriptBasedControllerCommandGenerator;
@@ -39,23 +40,22 @@ import us.ihmc.robotModels.FullHumanoidRobotModel;
 import us.ihmc.robotics.controllers.ControllerFailureException;
 import us.ihmc.robotics.dataStructures.registry.YoVariableRegistry;
 import us.ihmc.robotics.dataStructures.variable.YoVariable;
-import us.ihmc.robotics.geometry.BoundingBox3d;
 import us.ihmc.robotics.random.RandomGeometry;
 import us.ihmc.robotics.referenceFrames.ReferenceFrame;
 import us.ihmc.robotics.robotController.RobotController;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.robotSide.SideDependentList;
 import us.ihmc.robotics.screwTheory.InverseDynamicsCalculatorListener;
+import us.ihmc.simulationConstructionSetTools.bambooTools.BambooTools;
+import us.ihmc.simulationConstructionSetTools.simulationTesting.NothingChangedVerifier;
+import us.ihmc.simulationConstructionSetTools.util.environments.CommonAvatarEnvironmentInterface;
+import us.ihmc.simulationConstructionSetTools.util.environments.DefaultCommonAvatarEnvironment;
 import us.ihmc.simulationconstructionset.HumanoidFloatingRootJointRobot;
 import us.ihmc.simulationconstructionset.SimulationConstructionSet;
-import us.ihmc.simulationconstructionset.bambooTools.BambooTools;
-import us.ihmc.simulationconstructionset.bambooTools.SimulationTestingParameters;
 import us.ihmc.simulationconstructionset.simulatedSensors.WrenchCalculatorInterface;
-import us.ihmc.simulationconstructionset.util.environments.CommonAvatarEnvironmentInterface;
-import us.ihmc.simulationconstructionset.util.environments.DefaultCommonAvatarEnvironment;
 import us.ihmc.simulationconstructionset.util.simulationRunner.BlockingSimulationRunner;
 import us.ihmc.simulationconstructionset.util.simulationRunner.BlockingSimulationRunner.SimulationExceededMaximumTimeException;
-import us.ihmc.simulationconstructionset.util.simulationTesting.NothingChangedVerifier;
+import us.ihmc.simulationconstructionset.util.simulationTesting.SimulationTestingParameters;
 import us.ihmc.tools.thread.ThreadTools;
 
 public class DRCSimulationTestHelper
@@ -123,6 +123,18 @@ public class DRCSimulationTestHelper
                                   boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep, boolean automaticallySpawnSimulation,
                                   HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
    {
+      this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, drcNetworkModuleParameters,
+           highLevelBehaviorFactoryToAdd, initialSetup, addFootstepMessageGenerator, useHeadingAndVelocityScript, cheatWithGroundHeightAtForFootstep,
+           automaticallySpawnSimulation, walkingScriptParameters, true);
+   }
+   
+   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel,
+                                  DRCNetworkModuleParameters drcNetworkModuleParameters, HighLevelBehaviorFactory highLevelBehaviorFactoryToAdd,
+                                  DRCRobotInitialSetup<HumanoidFloatingRootJointRobot> initialSetup, boolean addFootstepMessageGenerator,
+                                  boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep, boolean automaticallySpawnSimulation,
+                                  HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters, boolean useBlockingSimulationRunner)
+   {
       this.controllerCommunicator = PacketCommunicator.createIntraprocessPacketCommunicator(NetworkPorts.CONTROLLER_PORT,
                                                                                             new IHMCCommunicationKryoNetClassList());
       this.testEnvironment = commonAvatarEnvironmentInterface;
@@ -178,8 +190,11 @@ public class DRCSimulationTestHelper
       scs = simulationStarter.getSimulationConstructionSet();
       sdfRobot = simulationStarter.getSDFRobot();
       avatarSimulation = simulationStarter.getAvatarSimulation();
-      blockingSimulationRunner = new BlockingSimulationRunner(scs, 60.0 * 10.0);
-      simulationStarter.attachControllerFailureListener(blockingSimulationRunner.createControllerFailureListener());
+      if (useBlockingSimulationRunner)
+      {
+         blockingSimulationRunner = new BlockingSimulationRunner(scs, 60.0 * 10.0);
+         simulationStarter.attachControllerFailureListener(blockingSimulationRunner.createControllerFailureListener());
+      }
 
       if (simulationTestingParameters.getCheckNothingChangedInSimulation())
       {
@@ -291,12 +306,18 @@ public class DRCSimulationTestHelper
 
    public void simulateAndBlock(double simulationTime) throws SimulationExceededMaximumTimeException, ControllerFailureException
    {
-      blockingSimulationRunner.simulateAndBlock(simulationTime);
+      if (blockingSimulationRunner != null)
+      {
+         blockingSimulationRunner.simulateAndBlock(simulationTime);
+      }
    }
 
    public void destroySimulation()
    {
-      blockingSimulationRunner.destroySimulation();
+      if (blockingSimulationRunner != null)
+      {
+         blockingSimulationRunner.destroySimulation();
+      }
       blockingSimulationRunner = null;
       if (avatarSimulation != null)
       {
@@ -382,16 +403,16 @@ public class DRCSimulationTestHelper
       scs.selectCamera("testCamera");
    }
 
-   public void assertRobotsRootJointIsInBoundingBox(BoundingBox3d boundingBox)
+   public void assertRobotsRootJointIsInBoundingBox(BoundingBox3D boundingBox)
    {
       assertRobotsRootJointIsInBoundingBox(boundingBox, getRobot());
    }
 
-   public static void assertRobotsRootJointIsInBoundingBox(BoundingBox3d boundingBox, HumanoidFloatingRootJointRobot robot)
+   public static void assertRobotsRootJointIsInBoundingBox(BoundingBox3D boundingBox, HumanoidFloatingRootJointRobot robot)
    {
       Point3D position = new Point3D();
       robot.getRootJoint().getPosition(position);
-      boolean inside = boundingBox.isInside(position);
+      boolean inside = boundingBox.isInsideInclusive(position);
       if (!inside)
       {
          fail("Joint was at " + position + ". Expecting it to be inside boundingBox " + boundingBox);
